requirements.txt
# MCP SDK
mcp[cli]>=1.0.0

# Vault file handling
python-frontmatter==1.0.1
pyyaml==6.0.1
watchdog>=3.0.0

# Configuration
python-dotenv==1.0.0
